Dispatcher Refactoring: De-lazify action config
Delegating to lazy properties has no benefit here. It was originally intended to
speedup Dispatcher construction, by reducing allocation & avoiding prefs lookup.
But Lazy objects themselves have to be allocated, and we have to do an extra
indirection to retrieve actual values later. So, it was a pre-mature optimization,
entirely unnecessary here.